voxelization correctly

+++ IlluminationComparison/SparseVoxelOctree.h
@@ -32,9 +32,9 @@ private:
 	void cpuVoxelListCapture(ID3D11Device* device, ID3D11DeviceContext* context);
 	void cpuOctreeCapture(ID3D11Device* device, ID3D11DeviceContext* context);
 
-	int worldWidth = 8;
-	int voxelCount = 0;
-	int	voxelDim = 64; // 256*256*256 + mip mapped octree for memory size
+	float  worldWidth = 8;
+	int    voxelCount = 0;
+	float  voxelDim = 64; // 256*256*256 + mip mapped octree for memory size
 	int octreeSize = 0;
 	
 	ID3D11Buffer			  *counter;
